# Scraping: improve memory handling of scrape buffers