Context here is that I am analyzing Prometheus scraping a range of targets which return from a few KB to tens of megabytes (kube-state-metrics), and a large number of buffers retained in the scape pool, roughly matching those sizes.

Improvements here are stylistic:
* `util.Pool` returns `[]byte` instead of `interface{}`.  This essentially reverts #3835: since this pool is only used in one place making it generic just complicates things.

And for efficiency:
* Add a pool for slice headers, similar to `zeropool`, saves 1 allocation of 24 bytes on every Put to the pool.  Just one for all the buckets.
* Fewer pool buckets. Previously we had sizes in bytes: 1000, 3000, 9e3, 27e3, 81e3, 243e3, 729e3, 2.2e6, 6.6e6, 1.97e6, 59e6.
Now they are 8000, 24000, 72e3, 216e3, 648e3, 1.9e6, 5.8e6 and "outsize".  So it takes less time to find the right bucket, and less space is wasted from rounding up to the maximum size of the bucket.
* We now pool objects bigger than the largest bucket, in an "outsize" pool.  This is a slice not a `sync.Pool` since those maintain a separate free-list for each CPU core.  Cleaning it out on each garbage-collection cycle is a little complicated.

In my ~8GB Prometheus this gives a ~300MB reduction.  It doesn't show up so much in the benchmarks.

<details><summary>Benchmarks</summary>
<p>
<pre>
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                               │    before    │               after               │
                                               │    sec/op    │   sec/op     vs base              │
ScrapeAndReport/#00-4                            188.6µ ± 30%   191.9µ ± 8%       ~ (p=0.818 n=6)
ScrapeAndReport/application/openmetrics-text-4   191.6µ ± 13%   186.5µ ± 4%       ~ (p=0.240 n=6)
geomean                                          190.1µ         189.2µ       -0.48%

                                               │    before     │                after                │
                                               │     B/op      │     B/op       vs base              │
ScrapeAndReport/#00-4                            36.19Ki ± 94%   36.14Ki ± 19%       ~ (p=0.065 n=6)
ScrapeAndReport/application/openmetrics-text-4   37.24Ki ±  0%   37.19Ki ±  0%  -0.12% (p=0.004 n=6)
geomean                                          36.71Ki         36.66Ki        -0.12%

                                               │   before    │               after               │
                                               │  allocs/op  │  allocs/op   vs base              │
ScrapeAndReport/#00-4                            844.0 ± 95%   842.0 ± 20%  -0.24% (p=0.015 n=6)
ScrapeAndReport/application/openmetrics-text-4   858.0 ±  0%   856.0 ±  0%  -0.23% (p=0.002 n=6)
geomean                                          851.0         849.0        -0.24%
<pre>
</p>
</details>