@bboreham 2023-10-13
In Prombench the memory reduction is clearest on the 'Next GC' line, with main consistently above this PR.
The peak since it started is 24.1GB vs 24.6GB:

<img width="441" alt="image" src="https://github.com/prometheus/prometheus/assets/8125524/e78f1471-8eaa-401a-a7db-5d52cd74518f">