@bboreham 2024-02-07
Going to close this as not worth the extra lines of code.